refactor(pro-vrf): restructure module
