+++ provider-verify/src/main/kotlin/io/github/wliamp/pro/vrf/config/VerifyProviderProperties.kt
@@ -3,7 +3,7 @@ package io.github.wliamp.pro.vrf.config
 import org.springframework.boot.context.properties.ConfigurationProperties
 
 @ConfigurationProperties(prefix = "provider.oauth")
-internal data class OauthProviderProperties(
+internal data class VerifyProviderProperties(
     var facebook: FacebookProps = FacebookProps(),
     var google: GoogleProps = GoogleProps(),
     var zalo: ZaloProps = ZaloProps()
