+++ provider-verify/src/main/kotlin/io/github/wliamp/pro/vrf/oauth/ZaloOauth.kt
@@ -1,13 +1,13 @@
-package io.github.wliamp.pro.vrf.data
+package io.github.wliamp.pro.vrf.oauth
 
-import io.github.wliamp.pro.vrf.config.OauthProviderProperties
+import io.github.wliamp.pro.vrf.config.VerifyProviderProperties
 import org.springframework.core.ParameterizedTypeReference
 import org.springframework.web.reactive.function.client.WebClient
 import reactor.core.publisher.Mono
 import kotlin.collections.get
 
 internal class ZaloOauth internal constructor(
-    private val props: OauthProviderProperties.ZaloProps,
+    private val props: VerifyProviderProperties.ZaloProps,
     private val webClient: WebClient
 ) : IOauth {
     private val provider = "zalo"
