+++ provider-verify/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.autoconfiguration.imports
@@ -1,1 +1,1 @@
-io.github.wliamp.pro.vrf.config.OauthProviderAutoConfig
+io.github.wliamp.pro.vrf.config.VerifyProviderAutoConfig
